improve: add get_confpath() to cinit object.
Not integrated yet to existing code.
Allow to get the path to the config file of a component.
For example, the config file of the webcore component is
defined by the environment variable CPS_CONFPATH_WEBCORE.
If the environment variable isn't set, then a hard-coded
default value will be used (in this case: ~/etc/webserver.conf)