fix(build): Package dependencies for the snmp engine were missing
We need libsmi and probably net-snmp-utils on CentOS, and on Debian
it was missing snmp-mibs-downloader.

(We don't care for Debian 8 anymore)

Since snmp is only a component of CAT, this should be in CAT's package
files, but oh well.  Current build is broken, and some of these dependencies
are already in Core packages.