fix(weather): rewrite get_ok_ko to fix performance issues
The go engine stat used to add a new point with fields ok and ko for
each event, with value 0 or 1.

It has been modified to only add fields with the value 1.

This allows to get the last ok or ko event with `last(ok)` or
`last(ko)`, which is a lot faster than using a query `where ko = 1`.

This commit adapts the service weather API to this change.