fix(build): Use --no-download --system-site-packages so that vens don't auto-upgrade pip
We can't install any version newer than 20.1.1, because later versions
have more and more problems with Python 2.

But, when you create a virtualenv, the latest version of pip is always
downloaded by default, unless you use --no-download. This is
only mentioned in the changelog of virtualenv 14.0.0:

    Download new releases of the preinstalled software from PyPI when
    there are new releases available. This behavior can be disabled
    using --no-download.

Since we always start with an (old) version of Pip from our build OS,
we can use that one, install a virtualenv which uses that old pip
version and doesn't upgrade it, and then call "pip install -U
pip==20.1.1" to upgrade it to the last "good" Python 2 version.

Phew.