feat(canoctl): Don't use notify handlers in ansible-role-redis
In Ansible's "design", a notify handler will not necessarily be executed
at least at the end of a role, it can also be triggered when it feels
like it, such as the very end of a whole playbook, so that Redis would
only be started after all Canopsis engines have started.

So, refute yet another Ansible "best practice" and just execute our
systemd commands in sequential order, instead of letting a notify
handler execute it in some random order for no added value.

(InfluxDB has the same "bug", but it doesn't cause us a bug; and
for Nginx it's not a problem either.)

Issue #2877.