feat(build): Always make a native Go build, don't enforce linux-amd64 there
It is not the responsibility of the local build system to enforce
this. At this stage, the format should be local and neutral.

The final pipeline producing the officially supported builds should
enforce linux-amd64, because that's what we validate in QA. But then
this responsibility should only appear in the meta-build-script
calling these Makefiles.

(Also use CGO_ENABLED=1 everywhere, instead of using a variable which
is always set to 1.)