fix(build): Repair the Pro/Community build even more
Use a WORKDIR that plays nicely with our go.mod trick, and make the
Pro makefiles a bit less stupid.

As a side effect, the Makefiles in Canopsis Pro are still as
repetitive as before, but they now only build 6 times instead of 36!

Also make the makefiles more verbose. Don't hide build things (at
least not without a VERBOSE flag).