fix(makefile): Replace OS's realpath --relative-to with GNU Make's realpath
The one from GNU Make is more portable, especially for non GNU/Linux
systems, such as macOS where there's no /usr/bin/realpath.

It makes all paths appear as absolute paths, but that's "good-enough",
especially for a new build system where we want everything to be
extremely explicit, so that build errors don't depend on moving
relative paths with all the context that it would require.