feat(makefile): Force a linux/amd64 platform build
The final product is always going to be a linux/amd64 container, so
force this when building Docker images, because this is what should
be tested everyday.

A linux/arm64 setup could theoretically work, but at the moment,
building a Golang binary on Alpine 3.12 arm64 just gives a
Segmentation fault.

Should help macOS M1 users a bit.