feat(swagger): Detach the swagger container from GOLANG_IMAGE_TAG
I don't see any problem having this part of the tree use its own
detached (and variable) version of Golang, since it's only used to
internally maintain the documentation.

Also try to move it to Alpine, so that building the documentation
requires a bit less space and time.

If this broke your setup: please ping me.