feat(ci): Drop the historical go-engines-community CI
This is FUBAR. It's been poorly maintained for years, it still relies
on the Python stack (which will be removed soon), it's full of "wait"
hacks, it's slow, etc. And now it doesn't work anymore (stuck in
waiting-for-MongoDB errors), blocking the new build system.

"make test" can still be run locally for now.

Work on a new CI system will start in the next Canopsis 4.5 sprint
(see for example the new !162 MR).