feat(build): Rename canopsis-api to canopsis-api-community and canopsis-api-pro
The new build system is dynamic, and so the various binaries in cmd/
must always have a unique name.

So, internally, we'll have distinct canopsis-api-community and
canopsis-api-pro (since they produce different binaries), but
"make install", the RPM packages and our containers will contain a
convenient /canopsis-api symlink for this.

Also make /canopsis.toml a dynamic symlink, while here.

Note to developers: DO NOT REVERT THIS CHANGE.