fix(build): Don't add binutils-gold to the base images
binutils-gold was a local test to see if the engines would build in
an arm64 layer (they do, but they segfault at runtime), but we don't
need it for the default x86_64 layers which are the only ones we
provide and support.

Remove binutils-gold to make sure it can't have any side effect.

Note: the one for swagger-init can be kept, since this one does work
on arm64, and it just locally updates the docs.