fix(build): Go back to Go 1.17.2
It was upgraded in MR !677, but this MR was done after the first RC,
so Go shouldn't be upgraded once we start the freezing period.

It also caused a fatal build error:

   docker.io/library/golang:1.17.4-alpine3.13: not found

because Go 1.17.4 bumped its Alpine version from 3.13 to 3.14, but
issue #3871 prevents us from upgrading the Alpine version in the
Canopsis 4.4 branch, anyway.