feat(deploy-ansible): Add our new postgresql-timescaledb role
Issue #4032.

The full Ansible playbook should be removed and replaced by proper
RPMs "at some point", but we aren't here yet, and as long as the
Python 2 is lying around, it's pretty difficult to inject proper
RPM dependencies in the main canopsis packages, and since TimescaleDB
itself has its dependency (PostgreSQL) that would be too complicated
for this sprint.