fix(build): Add an engine-che override to bring back the datasource plugin in Pro
The problem is that engine-che has the same name between Community and
Pro (although it was documented in #4036 that it shouldn't be done…)
so now Docker needs to have a different behaviour between a Community
and a Pro build (since the mongo.so file only exists on Pro), but
Dockerfile's COPY has no support for conditional keywords.  So the
only thing we have at our disposal is the "*" operator, and its side
effects: "*se*" will match "engine-service" on Community, and both
"engine-service" AND the "datasource" directory on Pro.  So,
datasource will be added purely by accident (yes), and just after
that we'll remove "engine-service" because we just took it to fool
the COPY operator.  I cried a little while having to do this.

Any use of conditions should be in the Makefile instead and we should
just call "make install" (and this was the plan), but since
conflicting engine names were still added, a shiny new
horrible-but-short workaround had to be found.

Hopefully, we'll remove it as soon as possible in the next sprint,
right?