Fully remove "migrate-metrics-meta" from Compose envs
As repeated by developers and reaffirmed in !2120, this program is not
intended for repeated runs. It is a one-time only job and is best out of
the Docker Compose services.

This change will also fix Compose envs which were broken for a few days,
because of commit f8da76b5 which deleted the definition for service
"migrate-metrics-meta", while still refering to it in all app services'
dependencies.