Handle error around rs.initiate() in mongodb-rs-init
With newer MongoDB 7.0, calling rs.initiate() when replicaset is already
initialized now results in an error.

Fortunately there is a precise error code for this case, allowing us to
correctly catch and ignore this error for the script to work after 1st
"up" of a given Canopsis Docker Compose stack.