fix(#5816): Add conditional rendering for author information
In `scenarios-list-expand-item.vue`, wrapped `v-flex` with a `v-if` condition to ensure the display of the scenario author only when the author exists. Similarly, in `event-filters-rule-summary.vue`, added a `v-if` condition to `c-information-block-row` to display the event filter author information only if the author is present. This ensures that the UI does not break when author details are missing.